+++ R/contrast.R
@@ -1,29 +1,30 @@
-#' Get contrast ratio of two colours
+#' Get Contrast Ratio of Two Colors
 #'
-#' Calculate the colour contrast ratio of two provided colours. Intended as an
-#' visual accessibility aid when selecting a text colour to place over a block-
-#' coloured background. The output value should be 4.5 or higher to ensure
+#' Calculate the color contrast ratio of two provided colors. Intended as an
+#' visual accessibility aid when selecting a text color to place over a single-
+#' color background. The output value should be 4.5 or higher to ensure
 #' sufficient contrast and readability. This function is currently based on
 #' \href{https://www.w3.org/TR/WCAG21/}{WCAG 2.1}. See
 #' \href{w3.org/TR/WCAG/#dfn-relative-luminance}{WCAG} for more information
 #' about the calculation.
 #'
-#' @param col1 Hex value preceded by '#' or a named colour.
-#' @param col2 Hex value preceded by '#' or a named colour.
+#' @param col1 Hex value preceded by '#' or a named color.
+#' @param col2 Hex value preceded by '#' or a named color.
 #' @param quiet Whether to print warning when the ratio value is lower than 4.5.
 #'
 #' @return A double.
 #' @export
 #'
-#' @examples get_ratio("#FFFFFF", "white")
+#' @examples cr_get_ratio("#FFFFFF", "white")
 
-get_ratio <- function(col1, col2, quiet = FALSE) {
+cr_get_ratio <- function(col_1, col_2, quiet = FALSE) {
 
-  d <- t(col2rgb(c(col1, col2))) / 255
+  d <- t(col2rgb(c(col_1, col_2))) / 255
 
   d <- apply(
-    d, 2,
-    function(x) ifelse(x <= 0.03928, x / 12.92, ((x + 0.055) / 1.055) ^ 2.4)
+    d, 2, function(x) ifelse(
+      x <= 0.03928, x / 12.92, ((x + 0.055) / 1.055) ^ 2.4
+    )
   )
 
   d <- as.data.frame(d)
@@ -39,5 +40,36 @@ get_ratio <- function(col1, col2, quiet = FALSE) {
   }
 
   return(cr)
+
+}
+
+#' Choose White or Black for Text Overlaying Supplied Color
+#'
+#' Given a background color, what's a better color for overlay text: white or
+#' black? Calculated as per \code{\link{cr_get_ratio}}. Defaults to black in the
+#' case of a tie.
+#'
+#' @param col_bg Hex value preceded by '#' or a named color.
+#'
+#' @return A character value: "white" or "black".
+#' @export
+#'
+#' @examples cr_choose_bw("grey90")
+
+cr_choose_bw <- function(col_bg) {
+
+  w <- cr_get_ratio(col_bg, "white", quiet = TRUE)
+  b <- cr_get_ratio(col_bg, "black", quiet = TRUE)
+
+  if (w > b) {
+    result <- "white"
+  } else if (w < b) {
+    result <- "black"
+  } else if (w == b) {
+    warning("There's a tie. Black chosen.")
+    result <- "black"
+  }
+
+  return(result)
 
 }