+++ NEWS.md
@@ -1,3 +1,11 @@
+# coloratio 0.0.0.9001
+
+* Added `cr_choose_bw()` function (#1).
+* Updated naming convention to `cr_*()` (#9).
+* Updated 'color' throughout (#11).
+* Added `cr_choose_bw()` to README.
+* Added code of conduct.
+
 # coloratio 0.0.0.9000
 
 * Added `get_ratio()` function.